bulduk-be/requirements.txt
django
djangorestframework
drf-yasg
django-cors-headers
djangorestframework-simplejwt
pymongo
django-ipware 